Mostly working movable GraphicsInterface

+++ src/LuaInterface.cpp
@@ -555,7 +555,7 @@ static int luaScreenClear(lua_State* L) {
 static int luaScreenRect(lua_State* L) {
   auto& luaInterface = getLuaInterfaceObject(L);
   auto [x, y, w, h, c] =
-    luaCheckedPullMultiple<int, int, int, int, int>(L, 1);
+    luaCheckedPullMultiple<float, float, float, float, int>(L, 1);
 
   luaInterface.graphicsInterface->rect(x, y, w, h, c);
   return 0;
@@ -564,7 +564,7 @@ static int luaScreenRect(lua_State* L) {
 static int luaScreenStrokeRect(lua_State* L) {
   auto& luaInterface = getLuaInterfaceObject(L);
   auto [x, y, w, h, c] =
-    luaCheckedPullMultiple<int, int, int, int, int>(L, 1);
+    luaCheckedPullMultiple<float, float, float, float, int>(L, 1);
 
   luaInterface.graphicsInterface->strokeRect(x, y, w, h, c);
   return 0;
@@ -573,7 +573,7 @@ static int luaScreenStrokeRect(lua_State* L) {
 static int luaScreenPoint(lua_State* L) {
   auto& luaInterface = getLuaInterfaceObject(L);
   auto [x, y, c] =
-    luaCheckedPullMultiple<int, int, int>(L, 1);
+    luaCheckedPullMultiple<float, float, int>(L, 1);
   luaInterface.graphicsInterface->point(x, y, c);
   return 0;
 }
@@ -581,7 +581,7 @@ static int luaScreenPoint(lua_State* L) {
 static int luaScreenText(lua_State* L) {
   auto& luaInterface = getLuaInterfaceObject(L);
   auto [x, y, str, fc, bc, align, baseline] =
-    luaCheckedPullMultiple<int, int, std::string_view,
+    luaCheckedPullMultiple<float, float, std::string_view,
                            int, int, pallet::GraphicsPosition,
                            pallet::GraphicsPosition>(L, 1);
   luaInterface.graphicsInterface->text(x, y, str, fc, bc, align, baseline);
